requirements.txt
# Dependencias necesarias para la ejecución en Streamlit Cloud
streamlit==1.50.0
pandas>=2.3.3
numpy>=2.0.0
plotly>=5.0.0
fpdf2>=2.7.7

# Dependencia para la base de datos DuckDB
duckdb>=0.10.2

# Dependencias para la SEGURIDAD (Inicio de Sesión)
streamlit-authenticator==0.3.3
pyyaml